Fix the format of COMP_E.mdt.xml files
The contents of eCatEnergyTemplate.xml did not match the format of the
old COMP_E.mdt.xml files (as seen by looking at a file produced by the
LabView software running on the North system). This confused the
ingestor and caused both the energy and throughput values to be stored
into a single channel (it was apparently random which one was stored).

This updates the template file to match the original format.